requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
asgiref==3.8.1
click==8.1.7
crispy-bootstrap4==2024.10
Django==5.1.3
django-crispy-forms==2.3
djangorestframework==3.15.2
fastapi==0.115.5
h11==0.14.0
idna==3.10
numpy==2.1.3
packaging==24.2
pandas==2.2.3
plotly==5.24.1
psycopg2-binary==2.9.10
pydantic==2.9.2
pydantic_core==2.23.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
six==1.16.0
sniffio==1.3.1
sqlparse==0.5.2
starlette==0.41.2
tenacity==9.0.0
typing_extensions==4.12.2
tzdata==2024.2
uvicorn==0.32.0
